API/get_product_by_ocassion_and_demographics/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-datastore==1.8.0